Create a new wiki page if one doesn’t exist for the uri
First a model is created, with name and slug based of the url visited.
Upon posting the model, The API returns the newly created object,
which also contains the appropriate permissions, created on the server-side.
Backbone automagically synchronises.

TODO: defer page creation to moment the first annotation is created (not easy)