Implement a relationship between time-based drivers and annotations
In views.js, MultiplexView, a view is defined that allows to register drivers:
a driver is a uri for which the Multiplex attempts to find a suitable Popcorn
instance. If the uri is the current page, it will be a basePlayer (abstract),
if it is an audio or video element it will be a standard Popcorn instance, etc.

PopCorn is an abstraction over HTML5 video, that allows to extend time based
logic to other elements, and is able to trigger events related to points in
the element’s timeline.

In views.js, AnnotationView, registerDriver, the view attempts to register
the annotation’s `about=` uri as a driver in the MultiplexView (by default,
this is the page).

In views.js, AnnotationView, updateAnnotationEvents, it is defined that
the annotation body looks for annotation elements, and registers them with
the associated driver through the aa plugin, using the data-begin and data-end
properties.

The AnnotationView all listen to an `aa:newDrivers` event that is fired by the
AnnotationCollectionView once all annotations are rendered: this triggers
updateAnnotationEvents.

In plugins/aa.popcorn.js it is defined that a passed element gets triggered
with `start` and `end` events when it arrives at these points in the video..

In global-events.js, listeningAnnotations we determine what needs to happen
when these start and end events fire: for now, it adds the class `active`
to the annotation.

An example: http://localhost:8000/pages/tests/
The two annotations to the left are linked to the video.