Change filter saving strategy: create an arborescence that maps to the url
Create a processed view, that serves static files, and if not found,
redirects to the process view.

Make the process view synchronous.
(Make it asynchronous again by passing a GET parameter async).

Make the process return a redirect to the resource.
(To make it return the info view, pass a GET parameter info)

Adapted a bit the front-end code, add an example of the pipeline
to the tests.