Versioning is coming. Youpi!
I was happy to find the "django-rcsfield" project.  Why didn't I find it
before? It is really nice and simple to version model fields, and has
many different backends like git, svn, bazaar etc.

But in our case what we wanted was to version models (specifically the
Page model) and not fields. I had to makes changes to it and instead
of just importing the library I merged it with the aawiki app. This is
how it works:

when an instance of a model suclassing RCSModel is saved, it is
serialized to JSON using tastypie, and committed to the repository.

Using the model manager one can retrieve older revision of the model,
basically the same way it is done in the API exemple here:

<https://code.google.com/p/django-rcsfield/wiki/QuickStart>

The API has been modified so one can request for a specific revision of
a page, for instance:

    http://localhost:8000/pages/api/v1/page/test/?rev=0f50eb8002

The resource returned from the server also has an extra field
"revisions" with the SHA1 of the commits featuring the resource, and the
annotations field is fully populated so we can just make one call the
Page API and retrieve a full document (a Page and its annotations). And
for that, I switched to backbone-relational to automatically create the
annotation collection.

Quick and dirty, more later.