Play and pause events now propagate to children: the timeline player works
When a driver fires `play` or `pause`, sections of annotations that refer to this driver are sought.
If these contain currently active annotations, these are sought for media elements.
the corresponding drivers are then also paused or played: it trickles down.

Caveat: a sound/movie might still be playing after its annotation is no longer active.
This media will not stop.

Caveat: for a video to be added to the timeline, it needs to be in a timed annotation.
For a stand-alone video-box, this is not intuitive yet (see the /pages/Tests/ for how it works)
we need some ui maybe.