Expand collapsed box on hover
This is a stop gap solution: the collapsed box will show the original box as well,
as you hover over it.
It would be better to do a solution where you have to click, and than the whole box
slides out: but that’s more involved an probably impossible in CSS alone.